pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["hayata-yamamoto <hayata.yamamoto.work@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
stanza = "^1.0.1"
grpcio_tools = "^1.29.0"
grpcio-tools = "^1.29.0"
grpcio_health_checking = "^1.29.0"
environs = "^7.4.0"
googleapis-common-protos = "^1.51.0"
grpcio_reflection = "^1.29.0"
grpcio = "^1.30.0"
protobuf = "^3.12.2"
mypy-protobuf = "^1.23"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
yapf = "^0.30.0"
pytest = "^5.4.2"
flake8 = "^3.8.1"
pytest-mock = "^3.1.0"
isort = "^4.3.21"
toml = "^0.10.1"
lxml = "^4.5.1"
overrides = "^3.0.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 119

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
